\name{upsSpikeIn}
\alias{upsSpikeIn}

\docType{data}
\title{Protein expression data from the CPTAC Technology Assessment (Study 6)}
\description{
	The data contains preprocessed expression values from CPTAC Technology Assessment,
	where equimolar concentrations of 48 human proteins (Sigma UPS 1) are spiked in
	five different concentrations into yeast proteome background and then processed.
	The example data contains only two groups from the study 6.
}

\usage{upsSpikeIn}
\format{
	\describe{
		\item{\code{upsSpikeIn}}{A numeric matrix of protein expressions.}
	}
}

\keyword{datasets}
